requirements.txt
# FastAPI and Server
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Configuration
pydantic-settings==2.6.1
pydantic==2.10.3

# Azure Monitor (includes OpenTelemetry dependencies)
azure-monitor-opentelemetry==1.8.2

# OpenTelemetry Instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-requests

# HTTP Client (for dependency tracking demo)
requests==2.32.3

# Testing
pytest==8.3.4
httpx==0.28.1
